refactor(api): move AI runtime contract into pkg/api
Move the provider runtime contract (Config, Provider/StreamingProvider, Response,
Event/EventKind/Event*, PermissionFunc/Request/Decision, the provider registry and
NewProvider) from pkg/ai into pkg/api so external consumers (clicky/aichat) can
depend on the stable api surface without importing the heavier pkg/ai runtime.
pkg/ai re-exports every moved symbol as a type alias / const / wrapper func, so all
existing call sites and pkg/ai/provider registration compile unchanged. Additive to
pkg/api (backwards compatible); the registry map stays unexported.
